requirements.txt
amqp==2.6.1
autopep8==1.5.7
billiard==3.6.4.0
cached-property==1.5.2
celery==4.4.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Flask==1.1.2
Flask-Cors==3.0.8
importlib-metadata==4.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==4.6.11
MarkupSafe==2.0.1
prompt-toolkit==3.0.18
pycodestyle==2.7.0
python-dotenv==0.17.1
pytz==2021.1
redis==3.5.3
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
vine==1.3.0
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.4.1
